+++ app/src/organisms/Desktop/Devices/RobotSettings/RobotSettingsComplianceReady/passwordinputfield.module.css
@@ -1,0 +1,8 @@
+.password_visibility_button {
+  display: flex;
+  padding: 0;
+  border: none;
+  background: transparent;
+  color: inherit;
+  cursor: pointer;
+}
