+++ app/src/organisms/Desktop/Devices/RobotSettings/RobotSettingsComplianceReady/personalaccountsettings.module.css
@@ -70,12 +70,3 @@
   justify-content: flex-end;
   gap: var(--spacing-8);
 }
-
-.password_visibility_button {
-  display: flex;
-  padding: 0;
-  border: none;
-  background: transparent;
-  color: inherit;
-  cursor: pointer;
-}
